Refactor the validation code - move to TypeScript in separate file
